add: MVC 패턴 폴더구조로 변경
