+++ controllers/detailController.js
@@ -1,0 +1,12 @@
+import {
+  allergieService,
+  allergieDrinkService,
+  categoryService,
+  drinkService,
+  imageService,
+  nutritionService,
+} from '../services';
+
+const detailController = () => {};
+
+export default detailController;
