+++ controllers/index.js
@@ -1,0 +1,9 @@
+import userController from './userController';
+import listController from './listController';
+import detailController from './detailController';
+
+export default {
+  userController,
+  listController,
+  detailController,
+};
