+++ controllers/listController.js
@@ -1,0 +1,5 @@
+import { categoryService, drinkService, imageService } from '../services';
+
+const listController = () => {};
+
+export default listController;
