+++ controllers/userController.js
@@ -1,0 +1,5 @@
+import { userService } from '../services';
+
+const userController = () => {};
+
+export default userController;
