+++ prisma/index.js
@@ -1,0 +1,5 @@
+import { PrismaClient } from './src/generated/client';
+
+const prisma = new PrismaClient();
+
+export default prisma;
