+++ routes/listRouter.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import { listController } from '../controllers';
+
+const router = express.Router();
+
+router.get('/list', listController);
+
+export default router;
