+++ services/imageService.js
@@ -1,0 +1,35 @@
+import prisma from '../prisma';
+import express from 'express';
+
+const app = express();
+
+const imageService = () => {
+  app.post('/images', async (req, res) => {
+    await prisma.$queryRaw`
+    INSERT INTO images 
+      (id, drink_id, image_url) 
+    VALUES 
+      (1,1,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/08/[9200000003661]_20210819094346176.jpg'),
+      (2,2,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000002487]_20210426091745467.jpg'),
+      (3,3,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000000479]_20210426091843897.jpg'),
+      (4,4,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000002081]_20210415133656839.jpg'),
+      (5,5,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000000487]_20210430112319040.jpg'),
+      (6,6,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/03/[9200000003509]_20210322093452399.jpg'),
+      (7,7,'https://image.istarbucks.co.kr/upload/store/skuimg/2020/09/[9200000002672]_20200921171223898.jpg'),
+      (8,8,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000000038]_20210430113202458.jpg'),
+      (9,9,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/02/[9200000001636]_20210225093600536.jpg'),
+      (10,10,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/04/[9200000003285]_20210416154437069.jpg'),
+      (11,11,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/02/[9200000001635]_20210225092236748.jpg'),
+      (12,12,'https://image.istarbucks.co.kr/upload/store/skuimg/2021/03/[9200000003507]_20210322093414289.jpg');
+  `;
+
+    const image = await prisma.$queryRaw`
+    SELECT *
+    FROM images;
+  `;
+
+    res.json(image);
+  });
+};
+
+export default imageService;
