+++ services/index.js
@@ -1,0 +1,17 @@
+import allergieDrinkService from './allergieDrinkService';
+import allergieService from './allergieService';
+import categoryService from './categoryService';
+import drinkService from './drinkService';
+import imageService from './imageService';
+import nutritionService from './nutritionService';
+import userService from './userService';
+
+export default {
+  allergieDrinkService,
+  allergieService,
+  categoryService,
+  drinkService,
+  imageService,
+  nutritionService,
+  userService,
+};
